Porting index_add to structured kernels, add an out variant
This PR attempts to port `index_add` to structured kernels, but does more than that:

* Adds an `out=` variant to `index_add`
* Revises `native_functions.yaml` registrations, to not have multiple entries and instead pass default value to `alpha`.
* Changes in `derivatives.yaml` file for autograd functioning

Follow-up PRs in near future will attempt to refactor the OpInfo test, and will give another look at tests in `test/test_torch.py` for this function. (hence the use of ghstack for this)

This is WIP because there are tests failing for `Dimname` variant on mobile/android builds, and I'm working on fixing them.

ghstack-source-id: f7b2e3fa9045a73746b36a25677fa1d447ae66e3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/65993